Move controller logic into a separate package
